add entity serialization and save system

+++ core/src/com/pignic/spacegrinder/component/ShipPart.java
@@ -4,13 +4,13 @@ import com.badlogic.ashley.core.Component;
 
 public class ShipPart implements Component {
 
-	private final int shipId;
+	private final long shipId;
 
-	public ShipPart(final int shipId) {
+	public ShipPart(final long shipId) {
 		this.shipId = shipId;
 	}
 
-	public int getShipId() {
+	public long getShipId() {
 		return shipId;
 	}
 }
